Allocation and deallocation functions
This patch adds allocation and deallocation functions for IP sets.
For each type of set, there are four new functions:

  • “init” — initializes an already-declared instance

  • “done” — frees any memory used to store the set internally, but
    not the “ipvX_set_t” instance itself

  • “new” — allocates a new set instance, and then initializes it

  • “free” — finalizes a set instance, and then frees it

The first two are intended for use with sets that you allocate
directly on the stack, whereas the last two are for sets that you have
to allocate on the heap.