Templated definitions
This patch lets us use templates to define the IPv4 and IPv6 set types
and functions, since there will be a lot of overlap.  (The only real
differences will be the names of the types and functions, and the
number of bits allowed in the values.)  Instead of lots of duplicate
code, we now have a single definition of each function, where the
parts that are IPv4- or IPv6-specific are references using macros.