Don't automatically add libraries to build paths
Before, when CheckLibInPath found a library, it would add any
necessary entries to CPPPATH, LIBPATH, and LIBS.  That's a bit
heavy-handed — for instance, once we add the unit tests, we only need
the test program to include libcheck library.  Now, CheckLibPath
creates <libname>_CPPPATH, etc., variables, that you have to manually
add to the dependencies of your targets.