Tests now run on Linux
The Linux linker doesn't record the location of the library when you
use “-lipset” to link against the library.  So when you try to run the
tests, the OS can't find the library to load in.  Since we're not
installing the test script, we can get around this by linking against
the library using the relative path to the .so file — that way, the
Linux linker can find the library without us having to set
LD_LIBRARY_PATH.