ipset_ipv[46]_memory_size function
This patch adds a function that returns the amount of memory used
to store an IP set.  This doesn't take into account the fact that
BDD nodes can be shared among multiple sets — so if you add
together the “memory_size” values for all of your sets, this is
probably much more than the actual total memory requirement.
Also, the buddy package hides its BddNode type, so we can't use
sizeof to calculate the size of each node.  Instead, we have to
hard-code this to 20 bytes.