Using CUDD instead of Buddy as BDD library
This patch updates the ipset library to use CUDD [1] instead of Buddy
as the underlying BDD library.  The main reason for this is that CUDD
supports “algebraic decision diagrams” (ADDs), where the value of the
binary function can be any integer or pointer, rather than just a 0 or
1.  We'll use this to implement IP maps, which map a range of IP
addresses to a value.