IP maps
This patch adds IP maps, which are like IP sets, but instead of just
having a particular address be in the set or not, each address is
mapped to some value.  The value can be a pointer or an integer — we
use the intptr_t type to ensure that the integer is the same size as a
pointer would be.  We use the “algebraic decision diagrams” from the
CUDD library to implement these, though we had to modify the CUDD
library to store intptr_t values instead of doubles.  A better
modification to the library would introduce IDDs — “integer” decision
diagrams — which would give us the integral values we need without
clobbering anyone else who relies on double-valued ADDs.