Don't preallocate BDD variables
It turns out that if we preallocate variables in our call to
Cudd_Init, these variables aren't used when we assign a variable to
each bit in an IPv4 or IPv6 address.  Since we only allocate the
variables once during the lifetime of a program, it's acceptable to
not preallocate any, and have CUDD resize the array as necessary as we
allocate the variables for each bit.