Merging v4 and v6 set and map types
Before, we had four main types in the ipset library: IPv4 and IPv6
versions each of a set type and a map type.  However, we're using the
same BDDs and ADDs under the hood to represent both the IPv4 and IPv6
versions.  We can merge those into a single set type and a single map
type; we just construct a BDD using the 32 IPv4 variables if we add an
IPv4 address, or using the 128 IPv6 variables if we add an IPv6
address.  This also lets us have a “mixed” set, which contains both
IPv4 and IPv6 addresses.  Sets and maps still need to be separate
types, since we can't mix ADDs and BDDs.