Storing IP sets and maps to disk
This patch adds functions for reading and writing IP sets and maps to
disk.  It uses the DDDMP library that is included with CUDD, so we
don't have any control over the output format.  In particular, DDDMP
only supports the “text” format for ADDs, so our map files might be
rather large.  Gzip will probably work well on them, though.