Size tests work on 32-bit machines
There are several test cases that verify the size of an IP set or map;
for these tests, the expected size was calculated on a 64-bit machine.
On 32-bit machines, the sizes will be half as big.  This patch updates
the test cases to use the correct expected value depending on the word
size of the current machine.