Integer constant macros with boost < 1.42
A recent Boost update defines the UINTX_C constant macros whenever
<boost/cstdint.cpp> is included.  In earlier versions, they were only
defined if __STDC_CONSTANT_MACROS is defined, matching the C99
standard.  Our test cases tried to define this macro accordingly, but
unfortunately, defined it too late.  This patch moves the definition
earlier in the files so that we have the constant macros with earlier
versions of Boost.