Cached function objects
We currently use a cache to memoize the results of creating terminal
and nonterminal nodes, to ensure that our BDDs are reduced.  Several
other functions will need to be cached in this way, so this patch
factors out that behavior into a separate class.  The new class uses a
slightly different interface; it caches the results of a unary
function.  We've defined unary functions for creating terminals and
nonterminals; new cached operators will also be implemented in terms
of unary functions like this.

The rest of the code was refactored slightly, as well.  We now have an
“engine” class which groups together the two caches we currently have;
it will include any further caches that we define, as well.