Evaluating a BDD
This patch adds an evaluation function to the node classes.  It takes
a vector of bools, which is used as the input to the BDD's function.
The evaluation function steps through the BDD tree, yielding a result
as soon as it hits a terminal node.  The function is templated; the
input value can be any class, as long as it provides a [] operator
that returns something that can be cast to a bool.