Refactored cached functions
This patch updates the cached function interface to be slightly easier
to use.  You no longer have to take care of boxing up the arguments to
the function in a tuple; this is taken care of for you by the
cached_unary_function_t, cached_binary_function_t, and
cached_ternary_function_t classes.

Also, the classes now follow the Curiously Recurring Template Pattern
— i.e., the function class that implements the underlying function
should be a subclass of the appropriate cached_X_function_t class.
This makes it easy to define recursive cached functions, since they
can call the cached version of itself directly, via the cached_call()
method.