Simplified non-template interface
This patch modifies the basic BDD node classes to no longer be
templated on the Value type.  Now, all BDDs map to integer values —
the range_t type, specifically.  Before, our only engine
specialization was for bool values, which can be transparently used as
range_t instances.  To support IP maps, which map to arbitrary values,
we'll have a separate class that maintains a cache of the values seen
so far; the BDD will then store integers that index into this cache.