IPv4 and IPv6 don't overlap in sets
This patch ensures that we use a distinct set of BDD variables for
IPv4 and IPv6 addresses.  If we don't do this, it's possible to add an
IPv6 /32 network and make it appear that an IPv4 address was added as
well, assuming that the IPv4 address is a prefix of the IPv6 network.