Serializing IP sets
This patch adds a method for writing IP sets (and their underlying
BDDs) to an output stream.  We write out the nonterminals using a
depth-first search, which ensures that all node references point to
nodes that appear earlier in the stream.  Part of the file format is
the length of the set's representation, which means we first have to
count the number of reachable nodes to calculate the right size.

There are thorough unit tests for the underlying BDD output routines;
for IP sets, the results are large enough that it's tedious to include
in the test case programs.  Once we have reading methods, we'll add
some round-trip tests that don't require encoding the specific file
format into the test code.