Changed variable assignments
The variable assignments that we were using before wouldn't properly
discriminate between IPv4 and IPv6 addresses — it was possible to add
an IPv4 address and make it look like an unrelated IPv6 address was in
the set.  Our variable assignments had separate sets of variables for
IPv4 bits and IPv6 bits.  However, when checking whether an IPv6
address was in the set, if we encountered BDD nodes referencing the
IPv4 addresses, we'd have to choose some value for them.  Before, it
was hard to reproduce the bug, since these values were arbitrary —
their values would depend on the four bytes of memory immediately
preceding the IPv6 address.  If the set contained that exact IPv4
address, it would look like the IPv6 address was in the set, too.  By
setting a fixed interpretation for the IPv4 variables (such as always
0), we could create a reproducible test case.

The solution is to use a discriminator variable, which takes the value
of TRUE for IPv4 addresses, and FALSE for IPv6 addresses.  This also
allows us to merge the two sets of variables together; we don't need a
separate bank of variables for IPv4 and IPv6.  Under the new scheme,
variable 0 is the discriminator.  IPv4 address bits are stored using
variables 1-32, while IPv6 address bits use variables 1-128.