BDD node iterators
The BDD node cache can now iterate through all of the assignments in a
BDD.  When the iterator is dereferenced, the result is a pair,
representing the domain and range of the Boolean function that the BDD
represents.  The first element of the pair is an instance of the
assignment_t class, giving the values that each Boolean variable needs
to take.  The second element is the terminal value that results when
the Boolean function is evaluated with that variable assignment.