Expanded assignments
The assignment_t class now contains an iterator that “expands” the
assignment: for each indeterminate variable, there will be an expanded
assignment where the variable is TRUE, and one where it is FALSE.
That gives 2^n expanded assignments, where n is the number of
indeterminate variables in the assignment.  Since assignments aren't
capped (there's an arbitrary number of indeterminate variables at the
end), the caller must pass in the number of variables that they care
about.  No variables past that will be given values in the expanded
assignment.