Parse IP address strings
The IP address classes now support being constructed from a character
string.  This tries to parse the string using the standard inet_pton
POSIX function.  The constructors throw an exception if the string has
an invalid format.  The generic ip_addr_t class has this new
constructor, too — it tries to parse the string first as IPv4, then as
IPv6, using the result of whichever succeeds.