BDD nodes
This patch adds an initial implementation of Binary Decision Diagram
nodes.  A BDD is a binary tree representing a boolean function — i.e.,
a function that maps a vector of boolean variables to some value.
Each internal (or “nonterminal”) node represents one of the variables
in the vector; it has a “low” child for when the variable is false,
and a “high” child for when the variable is true.  Each leaf (or
“terminal”) node represents an output value of the function.  To get
the value for a particular boolean input, you start at the root of the
BDD tree, and follow the child links based on the values in the input
vector.  This should lead you to a terminal node, representing the
output value for that input.

For efficiency, BDDs are typically “ordered”, meaning that as you
progress through a BDD tree, the indices of the nonterminal node
variables always increase.  They're also typically “reduced”, meaning
that there's only ever one physically node instance of a particular
set of values.  Reduced ordered BDDs (or ROBDDs) are nice because any
boolean function has a single BDD node that represents it, so it's
very easy to compare functions for equality.

This patch adds a basic types nonterminal nodes, as well as a type for
node IDs.  Node IDs are constructed so that we tell whether a node is
terminal or nonterminal just from the ID.  The value of a terminal
node is stored right in the node ID, rather than having a full struct
for it.

There is also a “node cache“ class which enforces the reduced
constraint by maintaining a cache of physical node instances.  The
ordered constraint will be enforced by our implementation of the BDD
operators.