Logical and (∧)
This patch adds an operator for calculating the logical and (∧) of two
BDDs.  This uses the standard dynamic programming implementation,
where we cache the results of each application, in addition to the
usual caching of the terminals and nonterminals that we create.

The basic and_operator class works for any kind of BDD value, assuming
that we can apply the standard & operator to the terminal values.