Memory size
This patch adds a function to return the amount of memory used by an
IP set.

The BDD test case now generates a large amount of logging output,
which brings out a bug in the Waf build scripts — it hangs if the
amount of output exceeds the size of the buffer in a pipe.  To get
around this for now, we now have Waf print out the output of the test
cases.