Serializing IP sets
This patch adds a method for writing BDDs to an output stream.  We
write out the nonterminals using a depth-first search, which ensures
that all node references point to nodes that appear earlier in the
stream.  Part of the file format is the length of the BDD
representation, which means we first have to count the number of
reachable nodes to calculate the right size.