Consisting load and save function names
To load in a BDD, we need access to a node cache, which we use to
create new BDD nodes as they're read in.  We don't technically need
access to the cache for saving a BDD, though, since everything can be
obtained directly from the node pointers.  However, it's better to
have the two functions have similar APIs, so I'm adding a node cache
parameter to the save function, too.