Generic IP address type
The new ipset_ip_t type is a simple discriminated union wrapper around
IPv4 and IPv6 addresses.  The content of the address is stored in the
“addr” field.  The “is_ipv4” field is the discriminator.  There are
constructors for creating ipset_ip_t instances from a raw IPv4
address, a raw IPv6 address, or from a human-readable string.  In
addition, there is a function to generate a human-readable string,
regardless of which type of address is present.

The ip_set_t and ip_map_t interfaces have also been updated to include
function that take in ipset_ip_t instances.  That said, the ipset_ip_t
type should be useful even if you don't use the IP set or map types.