Variable assignments
The new ipset_assignment_t type is used to store variable assignments,
which basically represent a path in a BDD from the root to a terminal.
The assignment is the particular value (TRUE or FALSE) that each
variable takes along that path.  This includes an “either” value for
variables that don't appear on the path — those variables can take
either value and you'll end up at the same terminal node.

This type will form the basis for an iterator that will loop through
the assignments in a BDD, which will then be further wrapped to give
us an iterator through all of the IP addresses in an IP set.