BDD node iterators
We can now iterate through all of the assignments in a BDD, using the
new ipset_bdd_iterator_t type.  The iterator contains two values of
interest, representing the domain and range of the Boolean function
that the BDD represents.  The first value is an instance of the
ipset_assignment_t type, giving the values that each Boolean variable
needs to take.  The second element is the terminal value that results
when the Boolean function is evaluated with that variable assignment.