Expanded assignments
We now have an iterator that “expands” a variable assignment: for each
EITHER variable, there will be an expanded assignment where the
variable is TRUE, and one where it is FALSE.  That gives 2^n expanded
assignments, where n is the number of EITHER variables in the
assignment.  Since assignments aren't capped (there's an arbitrary
number of EITHER variables at the end), the caller must pass in the
number of variables that they care about.  No variables past that will
be given values in the expanded assignment.  (Even ones that aren't
EITHER.)