Initialize extra space in expanded assignment
The expanded assignments are stored as simple bit vectors — basically
a long, arbitrary-length integer, with each variable's value stored as
a single bit.  If the number of variables isn't a multiple of 8, then
there will be extra unused bits leftover in the last byte of the bit
vector.  This was causing problems with some of our test cases, that
used a simple memcmp() to compare the full contents of the bit vector,
including the unused bits at the end.  The code now initializes these
unused bits to 0, so that a memcmp() comparison works as expected.