Refactored BDD save functions
The internal functions that are used to save a BDD have been
refactored to be more generic.  Now, the functions that trace through
the nodes of the BDD has been separated from the functions that
actually output the data.  This will make it easier to support other
output formats, without having to reimplement as much logic.