Double-expansion of certain assignments
When iterating through an IP set, we use a BDD iterator to retrieve
all of the assignments in the BDD, and expand each of those
assignments to create individual IP addresses or networks.  Each
expansion step can produce either IPv4 addresses, or IPv6 addresses,
but not both.  Variable 0 tells us which kind of address to use.

However, it's possible for an assignment to give variable 0 a value of
EITHER, which indicates that the assignment contains both IPv4 and
IPv6 addresses.  (One trivial example of this is the empty set; its
BDD iterator yields a single assignment where all variables are
EITHER.)  Each expansion step can still only produce either IPv4 or
IPv6, so for this assignment, we have to expand it twice: once where
we've artificially set variable 0 to TRUE (IPv4), and once where we've
set it to FALSE (IPv6).

The IP set iterator now handles this case correctly.  To support this,
we had to add some additional state to the iterator to remember
whether the current assignment needs to be expanded twice, and if so,
which expansion we're currently in.